Bug Fix
Fixed a bug where after dragging and dropping an item, the slot still retains the item in its script. Clicking on that slot would transfer the item from another slot into the held item. E.G money was dragged from slot 1 to slot 3, clicking on slot 1 would, independent of the items location, transfer it to the held item slot